CHROMIUM: LSM: check vfsmount to avoid disengagement
Instead of locking the dentry associated with the internal mount
structure, use the vfsmount structure. This means we don't have to
check mount counts (in the face of bind mounts).

Additionally add checking of the actual backing device being pinned
instead of going arbitrarily after the rootdev, which may not be the
pinning origin (in the case of recovery kernels, etc).

BUG=b:21762937
TEST=gizmo build, umount a bind mount doesn't disable module pinning

Change-Id: I6c4e4e985996698b24abf0edc0c2873e327258d1
Signed-off-by: Kees Cook <keescook@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/277331
Reviewed-by: Filipe Brandenburger <filbranden@chromium.org>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Reviewed-by: Aaron Durbin <adurbin@chromium.org>