drm/tegra: Fix up dpms in driver
This patch fixes up a few nits in dpms such that the
connectors can turn off/on successfully.

In dc, we do a full disable in prepare() to ensure the
dc is turned off on dpms Off. We also force a full
modeset if we're transitioning from off to on in
atomic_check.

In dsi, we do a full disable in prepare() for the
same reasons as above. We also switch over from the
stub dpms function to using the atomic helper.

Change-Id: Ia4f39652c79faa36f4fde6799a97cc26ce3c4412
Signed-off-by: Sean Paul <seanpaul@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/281352
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Commit-Queue: Sean Paul <seanpaul@google.com>
Tested-by: Sean Paul <seanpaul@google.com>