CHROMIUM: tegra/sysedp: fix reading battery Imax
The Imax read from battery is in uA, but the driver
use it in mA, so it will cause the wrong on power
calculation.
This patch fix it, and add comments to clarify the
unit of the values reading from the battery.

BUG=chrome-os-partner:39762
TEST=cat /sys/kernel/debug/systemp_edp/status, the power
budget is expected.

Change-Id: I96e351b37a8a108f0d9c7f7c801c0ca9513c1983
Signed-off-by: Wei Ni <wni@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/282170
Tested-by: Wei Ni <wni.nvidia@gmail.com>
Reviewed-by: David Riley <davidriley@chromium.org>
Commit-Queue: Wei Ni <wni.nvidia@gmail.com>