CHROMIUM: regulator: max8973: rewrite enable/disable of output with external control
Rewrite the enable/disable control of the external control
to support MAX8973 as well as multiple version of MAX77621.

On MAX77621, some version EN pin act as SHUTDOWN and on
some version it act as EN pin.

This patch fixes the issue that the GPU will hang if
resuming from rail gated.

http://patchwork.ozlabs.org/patch/482239/

BUG=chrome-os-partner:41247
TEST=GPU can be rail gated and resumed without hang

Change-Id: I2719bf88217a54c6b324d49c734b922ca5425f00
Signed-off-by: Kary Jin <karyj@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/282682
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>
Commit-Queue: Joseph Lo <josephl@nvidia.com>
Tested-by: Joseph Lo <josephl@nvidia.com>