FROMLIST: i2c: tegra: update CONFIG_LOAD for new conifiguration
Once the new configuration is set on the conifg register of
I2C controller, it is require to update the CONFIG_LOAD register
to transfer the new SW configuration to actual HW internal
registers that would be used in the actual logic.

It is like, SW is programming only shadow registers through
regular configuration and when these load_config bit fields
are set to 1, it causes the regular/shadows registers
configuration transferred to the HW internal active registers.
So SW has to set these bit fields at the end of all regular
registers configuration. And these config_load bits are HW
auto-clear bits. HW clears these bit fields once the register
configuration is moved to HW internal active registers. So SW
has to wait until these bits are auto-cleared before going
for any further programming

This mechanism is supported on T124 and after this SoCs.

This is based on change done by
	Chaitanya Bandi <bandik@nvidia.com>

(git am from https://patchwork.ozlabs.org/patch/489643/)

BUG=chrome-os-partner:40280
TEST=Build and boot.

Change-Id: I0e16446482b75c559ea13135d8793ce80f89245c
Signed-off-by: Laxman Dewangan <ldewangan@nvidia.com>
Signed-off-by: Chaitanya Bandi <bandik@nvidia.com>
Signed-off-by: Andrew Chen <andrewc@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/282906
Reviewed-by: Laxman Dewangan <dewangan.laxman@gmail.com>
Reviewed-by: Benson Leung <bleung@chromium.org>