CHROMIUM: usb: xhci-tegra: Support dual-role operation
An optional extcon device may be specified which can be used to
determine the state of the dual-role port (port 0).

When switching to host-mode, disable any VBUS/ID overrides on the
dual-role port and perform the reset SSPI sequence in order to
re-enable SuperSpeed operation.  An autosuspend delay is added
since it may take several 100s of milliseconds for the SuperSpeed
device to be enumerated.  VBUS/ID overrides must also be cleared
at boot and at resume time if in host mode.

When switching out of host-mode, force a runtime resume so that
PMC control of the dual-role port and any wakeup events may be
disabled.

BUG=chrome-os-partner:41244
TEST=USB runtime PM and xHCI/XUDC role-switching on Smaug.

Change-Id: Ide4a5b6dad2d44c0fe3b8925fe670d9ff7068a40
Signed-off-by: Andrew Bresticker <abrestic@chromium.org>
Signed-off-by: Mark Kuo <mkuo@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/285299